+++ src/main/java/org/bytedeco/javacpp/Pointer.java
@@ -30,6 +30,7 @@ import java.lang.reflect.Modifier;
 import java.nio.Buffer;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
+import org.bytedeco.javacpp.annotation.Platform;
 import org.bytedeco.javacpp.tools.Generator;
 import org.bytedeco.javacpp.tools.Logger;
 
@@ -58,6 +59,7 @@ import org.bytedeco.javacpp.tools.Logger;
  *
  * @author Samuel Audet
  */
+@Platform
 public class Pointer implements AutoCloseable {
     /** Default constructor that does nothing. */
     public Pointer() {}