+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Output compiled libraries to user specified class path by default for input classes inside JAR files, etc
  * Add `HalfIndexer` to access `short` arrays as half-precision floating point numbers
 
 ### July 8, 2016 version 1.2.2